requirements.txt
websockets>=12.0
aiohttp>=3.9.0
scipy>=1.11.0
loguru>=0.7.2
python-dotenv>=1.0.0
numpy>=1.26.0
